Muestra de fitness con ints en el debug del final de ejecucion

+++ Práctica 2/G01P2/src/es/ucm/fdi/pev/AGenetico.java
@@ -151,7 +151,7 @@ public class AGenetico
 		}
 		textoMejorAbs = textoMejorAbs + "]";
 		System.out.println(textoMejorAbs);
-		System.out.printf("Mejor fitness: %f\n", mejor_fitness);
+		System.out.printf("Mejor fitness: %d\n", (int) mejor_fitness);
 		System.out.printf("Generacion del mejor: %d\n", generacionMejor);
 		// Obtener peor cromosoma
 		genes = peor_abs.getGenes();
@@ -164,7 +164,7 @@ public class AGenetico
 		}
 		textoPeorAbs = textoPeorAbs + "]";
 		System.out.println(textoPeorAbs);
-		System.out.printf("Peor fitness: %f\n", peor_fitness);
+		System.out.printf("Peor fitness: %d\n", (int) peor_fitness);
 		System.out.printf("Generacion del peor: %d\n", generacionPeor);
 		_grafica.dibujaGrafica();
 	}
